+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.12.0-rc.13",
+  sdkVersion: "3.12.0-rc.15",
   genVersion: "2.687.1",
-  userAgent: "speakeasy-sdk/typescript 3.12.0-rc.13 2.687.1 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 3.12.0-rc.15 2.687.1 2.0 @orq-ai/node",
 } as const;