+++ packages/orq-rc/RELEASES.md
@@ -3199,3 +3199,13 @@ Based on:
 - [typescript v3.12.0-rc.13] packages/orq-rc
 ### Releases
 - [NPM v3.12.0-rc.13] https://www.npmjs.com/package/@orq-ai/node/v/3.12.0-rc.13 - packages/orq-rc
+
+## 2025-08-27 05:57:02
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.606.5 (2.687.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.12.0-rc.15] packages/orq-rc
+### Releases
+- [NPM v3.12.0-rc.15] https://www.npmjs.com/package/@orq-ai/node/v/3.12.0-rc.15 - packages/orq-rc